blog_project_with_flowbite/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.3
django-cloudinary-storage==0.3.0
django-debug-toolbar==4.4.6
django-ratelimit==4.1.0
django-redis==5.4.0
django-user-agents==0.4.0
django-user-sessions==2.0.0
Faker==30.6.0
frozenlist==1.5.0
idna==3.10
kombu==5.4.2
maxminddb==2.6.2
multidict==6.1.0
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
setuptools==75.3.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==0.18.0
urllib3==2.2.3
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.17.1
